--- website/src/app/views/about/members.js
@@ -1,966 +1,0 @@
-import ethan from '../../../images/about/memberImages/ethan.jpg';
-import mingjia from '../../../images/about/memberImages/mingjia.jpg';
-import sonali from '../../../images/about/memberImages/sonali.jpg';
-import kamille from '../../../images/about/memberImages/kamille.jpg';
-import joshualiu from '../../../images/about/memberImages/joshualiu.jpg';
-import megha from '../../../images/about/memberImages/megha.jpeg';
-import xuan from '../../../images/about/memberImages/xuan.jpg';
-import joshuachan from '../../../images/about/memberImages/joshuachan.jpg';
-import phillip from '../../../images/about/memberImages/phillip.jpg';
-import zubair from '../../../images/about/memberImages/zubair.jpg';
-import ashley from '../../../images/about/memberImages/ashley.jpg';
-import alec from '../../../images/about/memberImages/alec.jpg';
-import claude from '../../../images/about/memberImages/claude.jpg';
-import alan from '../../../images/about/memberImages/alan.jpg';
-import sydney from '../../../images/about/memberImages/sydney.jpg';
-import rostam from '../../../images/about/memberImages/rostam.jpeg';
-import hang from '../../../images/about/memberImages/hang.jpg';
-import chase from '../../../images/about/memberImages/chase.jpg';
-import rashmi from '../../../images/about/memberImages/rashmi.jpg';
-import duron from '../../../images/about/memberImages/duron.jpg';
-import jenny from '../../../images/about/memberImages/jenny.jpg';
-import henry from '../../../images/about/memberImages/henry.jpg';
-import chris from '../../../images/about/memberImages/chris.jpg';
-import preston from '../../../images/about/memberImages/preston.jpg';
-import kasey from '../../../images/about/memberImages/kasey.jpg';
-import madhusudhan from '../../../images/about/memberImages/madhusudhan.jpg';
-import maxim from '../../../images/about/memberImages/maxim.jpeg';
-import sylvester from '../../../images/about/memberImages/sylvester.jpg';
-import jane from '../../../images/about/memberImages/jane.jpg';
-import zoya from '../../../images/about/memberImages/zoya.jpg';
-import jules from '../../../images/about/memberImages/jules.JPG';
-import alvin from '../../../images/about/memberImages/alvin.JPG';
-import allen from '../../../images/about/memberImages/allen.jpg';
-import cyre from '../../../images/about/memberImages/cyre.jpg';
-import anikait from '../../../images/about/memberImages/anikait.jpg';
-import becky from '../../../images/about/memberImages/becky.jpg';
-import angeline from '../../../images/about/memberImages/angeline.jpg';
-import christine from '../../../images/about/memberImages/christine.jpg';
-import hannah from '../../../images/about/memberImages/hannah.jpg';
-import lancy from '../../../images/about/memberImages/lancy.JPG';
-import meghna from '../../../images/about/memberImages/meghna.jpg';
-import meredith from '../../../images/about/memberImages/meredith.jpg';
-import michelle from '../../../images/about/memberImages/michelle.jpg';
-import sabrina from '../../../images/about/memberImages/sabrina.jpg';
-import vani from '../../../images/about/memberImages/vani.jpg';
-import vivian from '../../../images/about/memberImages/vivian.jpg';
-import sam from '../../../images/about/memberImages/sam.jpg';
-import rileychampion from '../../../images/about/memberImages/rileychampion.jpg';
-import rileywong from '../../../images/about/memberImages/rileywong.jpg';
-import michelleblin from '../../../images/about/memberImages/michelleblin.jpg';
-import mia from '../../../images/about/memberImages/mia.jpg';
-import matthew from '../../../images/about/memberImages/matthew.jpg';
-import michellelin from '../../../images/about/memberImages/michellelin.jpg';
-import kei from '../../../images/about/memberImages/kei.jpg';
-import justin from '../../../images/about/memberImages/justin.jpg';
-import crystal from '../../../images/about/memberImages/crystal.jpg';
-import thomas from '../../../images/about/memberImages/thomas.jpg';
-import noah from '../../../images/about/memberImages/noah.jpeg';
-import ted from '../../../images/about/memberImages/ted.jpg';
-import michaelpien from '../../../images/about/memberImages/michaelpien.jpg';
-import phil from '../../../images/about/memberImages/phil.jpg';
-import kory from '../../../images/about/memberImages/kory.jpg';
-import parsa from '../../../images/about/memberImages/parsa.jpg';
-import anna from '../../../images/about/memberImages/anna.jpeg';
-import charles from '../../../images/about/memberImages/charles.jpg';
-import rongmu from '../../../images/about/memberImages/rongmu.jpg';
-import haydenp from '../../../images/about/memberImages/haydenp.jpg';
-import keagan from '../../../images/about/memberImages/keagan.jpg';
-import melayna from '../../../images/about/memberImages/melayna.jpg';
-import pham from '../../../images/about/memberImages/pham.jpg';
-import meryl from '../../../images/about/memberImages/meryl.jpg';
-import orvielalzate from '../../../images/about/memberImages/orvielalzate.jpg';
-import tony from '../../../images/about/memberImages/tonyliu.jpg';
-import austin from '../../../images/about/memberImages/austin.jpg';
-import olivia from '../../../images/about/memberImages/olivia.JPG';
-import rochelle from '../../../images/about/memberImages/rochelle.jpg';
-import hannahc from '../../../images/about/memberImages/hannahc.jpg';
-import rayan from '../../../images/about/memberImages/rayan.jpg';
-import daniel from '../../../images/about/memberImages/daniel.jpg';
-import kyle from '../../../images/about/memberImages/kyle.jpg';
-import celine from '../../../images/about/memberImages/celine.jpg';
-import dang from '../../../images/about/memberImages/dang.jpg';
-import dingyi from '../../../images/about/memberImages/dingyi.jpg';
-import rachel from '../../../images/about/memberImages/rachel.jpg';
-import david from '../../../images/about/memberImages/david.jpg';
-import thien from '../../../images/about/memberImages/Thien.jpg';
-import joshuamontellano from '../../../images/about/memberImages/joshuamontellano.jpg';
-import megan from '../../../images/about/memberImages/megan.jpg';
-import sravantallapaka from '../../../images/about/memberImages/sravantallapaka.jpg';
-import mignonaprillee from '../../../images/about/memberImages/mignonaprillee.jpg';
-import andrea from '../../../images/about/memberImages/andrea.jpg';
-import jackie from '../../../images/about/memberImages/jackie.jpg';
-import rithvikT from '../../../images/about/memberImages/rithvikT.jpg';
-import lee from '../../../images/about/memberImages/lee.jpg';
-import alexandria from '../../../images/about/memberImages/alexandria.jpg';
-import allisonliu from '../../../images/about/memberImages/allison.jpg';
-import alyssa from '../../../images/about/memberImages/alyssa.png';
-import ashleyyeh from '../../../images/about/memberImages/ashley.png';
-import carlos from '../../../images/about/memberImages/carlos.jpg';
-import christina from '../../../images/about/memberImages/christinavu.jpg';
-import erica from '../../../images/about/memberImages/ericayi.jpg';
-import fiona from '../../../images/about/memberImages/fionawu.jpg';
-import isabel from '../../../images/about/memberImages/isabelpham.png';
-import ivyn from '../../../images/about/memberImages/ivyn.jpg';
-import jacqueline from '../../../images/about/memberImages/jacquelinele.jpeg';
-import jade from '../../../images/about/memberImages/jadeluu.JPG';
-import james from '../../../images/about/memberImages/jamesmolamphy.jpg';
-import juhi from '../../../images/about/memberImages/juhipatel.jpg';
-import justine from '../../../images/about/memberImages/justinechou.jpg';
-import kaila from '../../../images/about/memberImages/kailalong.jpeg';
-import maireen from '../../../images/about/memberImages/maireenespiritu.jpg';
-import minh from '../../../images/about/memberImages/minhnguyen.jpg';
-import serena from '../../../images/about/memberImages/serenatrang.jpg';
-import zoyasoy from '../../../images/about/memberImages/zoyasoy.png';
-import arzoo from '../../../images/about/memberImages/arzoo.jpg';
-import kaitlyn from '../../../images/about/memberImages/kaitlynsuzuki.jpg';
-import johnny from '../../../images/about/memberImages/johnny.jpg';
-import ethanso from '../../../images/about/memberImages/ethanso.jpg';
-import jared from '../../../images/about/memberImages/jared.jpg';
-import ruslan from '../../../images/about/memberImages/ruslan.jpg';
-import derek from '../../../images/about/memberImages/derek.jpg';
-import kenny from '../../../images/about/memberImages/kenny.jpg';
-import duongvu from '../../../images/about/memberImages/duongvu.jpg';
-import evan from '../../../images/about/memberImages/evan.JPG';
-import katie from '../../../images/about/memberImages/katie.jpg';
-import kavan from '../../../images/about/memberImages/kavan.jpeg';
-import qiranwang from '../../../images/about/memberImages/qiranwang.png';
-import caroline from '../../../images/about/memberImages/caroline.jpeg';
-import kyledoliente from '../../../images/about/memberImages/kyledoliente.jpg';
-import kanu from '../../../images/about/memberImages/kanu.jpeg';
-import mazenghosn from '../../../images/about/memberImages/mazenghosn.jpg';
-import hassan from '../../../images/about/memberImages/hassan.jpg';
-import edwardwu from '../../../images/about/memberImages/edwardwu.jpeg';
-import ryancho from '../../../images/about/memberImages/ryancho.jpg';
-import kojihirai from '../../../images/about/memberImages/kojihirai.jpg';
-import charleswu from '../../../images/about/memberImages/charleswu.jpg';
-import gurneetcheema from '../../../images/about/memberImages/gurneetcheema.jpeg';
-
-const members = {
-  info: [
-    {
-      name: 'Ethan Nguyen',
-      position: 'Co-President',
-      linkedinURL: 'https://www.linkedin.com/in/nguyenethan01/',
-      imageURL: ethan,
-    },
-    {
-      name: 'Chase Carnaroli',
-      position: 'Co-President',
-      linkedinURL: 'https://www.linkedin.com/in/chasecarnaroli/',
-      imageURL: chase,
-    },
-    {
-      name: 'Joshua Liu',
-      position: 'Internal VP',
-      linkedinURL: 'https://www.linkedin.com/in/joshua-liu/',
-      imageURL: joshualiu,
-    },
-    {
-      name: 'Sydney Chiang',
-      position: 'External VP',
-      linkedinURL: 'https://www.linkedin.com/in/sydneychiang/',
-      imageURL: sydney,
-    },
-    {
-      name: 'Mingjia Wang',
-      position: 'Executive Advisor',
-      linkedinURL: 'https://www.linkedin.com/in/mingjia-wang/',
-      imageURL: mingjia,
-    },
-    {
-      name: 'Sonali Chellappa',
-      position: 'VP of Content Creation',
-      linkedinURL: 'https://www.linkedin.com/in/sonali-chellappa-1a16aa166/',
-      imageURL: sonali,
-    },
-    {
-      name: 'Rostam Vakhshoori',
-      position: 'Co-VP of Projects',
-      linkedinURL: 'https://www.linkedin.com/in/rostamvakhshoori/',
-      imageURL: rostam,
-    },
-    {
-      name: 'Zubair Sidhu',
-      position: 'Co-VP of Projects',
-      linkedinURL: 'https://www.linkedin.com/in/zubair-sidhu/',
-      imageURL: zubair,
-    },
-    {
-      name: 'Kamille Suayan',
-      position: 'VP of Marketing',
-      linkedinURL: 'https://www.linkedin.com/in/kamille-suayan-2a8980b7/',
-      imageURL: kamille,
-    },
-    {
-      name: 'Vivian Chu',
-      position: 'Co-VP of Design',
-      linkedinURL: 'https://www.linkedin.com/in/vivianchu204/',
-      imageURL: vivian,
-    },
-    {
-      name: 'Meghna Kaligotla',
-      position: 'Co-VP of Design',
-      linkedinURL: 'https://www.linkedin.com/in/meghna-kaligotla-605/',
-      imageURL: meghna,
-    },
-    {
-      name: 'Jane Vo',
-      position: 'VP of Recruitment, Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/janehv/',
-      imageURL: jane,
-    },
-    {
-      name: 'Duron Gardezi',
-      position: 'Treasurer, Developer',
-      linkedinURL: 'https://www.linkedin.com/in/durongardezi/',
-      imageURL: duron,
-    },
-    {
-      name: 'Alan Chang',
-      position: 'VP of Outreach, Developer',
-      linkedinURL: 'https://www.linkedin.com/in/acxyz/',
-      imageURL: alan,
-    },
-    {
-      name: 'Alexandria Wang',
-      position: 'Design Lead',
-      linkedinURL: 'https://www.linkedin.com/in/alexandria-wang-44066a187/',
-      imageURL: alexandria,
-    },
-    {
-      name: 'Allen Luo',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/allenxluo',
-      imageURL: allen,
-    },
-    {
-      name: 'Anikait Rao',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/anikait-rao/',
-      imageURL: anikait,
-    },
-    {
-      name: 'Chris Tian',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/cwtian/',
-      imageURL: chris,
-    },
-    {
-      name: 'Henry Gip',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/henry--gip/',
-      imageURL: henry,
-    },
-    {
-      name: 'Justine Chou',
-      position: 'Design Lead',
-      linkedinURL: 'https://www.linkedin.com/in/justine-chou/',
-      imageURL: justine,
-    },
-    {
-      name: 'Megha Kak',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/megha-kak/',
-      imageURL: megha,
-    },
-    {
-      name: 'Meredith Quintana',
-      position: 'Design Lead',
-      linkedinURL: 'https://www.linkedin.com/in/meredith-quintana-95880619b/',
-      imageURL: meredith,
-    },
-    {
-      name: 'Rashmi Sharma',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/rashmi-sharma-432b31178/',
-      imageURL: rashmi,
-    },
-    {
-      name: 'Sabrina Nguyen Hoang',
-      position: 'Design Lead',
-      linkedinURL: 'https://www.linkedin.com/in/sabrina-nguyen-hoang/',
-      imageURL: sabrina,
-    },
-    {
-      name: 'Sylvester Yue',
-      position: 'Tech Lead',
-      linkedinURL: 'https://www.linkedin.com/in/sylvester-yue/',
-      imageURL: sylvester,
-    },
-    {
-      name: 'Alvin Chen',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/alvinchenn',
-      imageURL: alvin,
-    },
-    {
-      name: 'Andrea AhSue',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/andrea-ahsue',
-      imageURL: andrea,
-    },
-    {
-      name: 'Anna Chan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/anna-channa',
-      imageURL: anna,
-    },
-    {
-      name: 'Arzoo Singh',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/arzoosingh/',
-      imageURL: arzoo,
-    },
-    {
-      name: 'Ashley Liu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/ashley-liu-6314731b9/',
-      imageURL: ashley,
-    },
-    {
-      name: 'Austin Truong',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/austin-truong-b32b68165/',
-      imageURL: austin,
-    },
-    {
-      name: 'Caroline Wang',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/caroline-wang-55b3711b8',
-      imageURL: caroline,
-    },
-    {
-      name: 'Celine Kim',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/sellin-celine-kim-5303271b4/',
-      imageURL: celine,
-    },
-    {
-      name: 'Charles Lee',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/charles-min-lee-978834214/',
-      imageURL: charles,
-    },
-    {
-      name: 'Charles Wu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/charles-wu-b2ba09225/',
-      imageURL: charleswu,
-    },
-    {
-      name: 'Christine Nguyen',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/christine-nguyen-89abb4172/',
-      imageURL: christine,
-    },
-    {
-      name: 'Claude Yan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/claude-yan-b41a3216a/',
-      imageURL: claude,
-    },
-    {
-      name: 'Crystal Lee',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/crystalleesf/',
-      imageURL: crystal,
-    },
-    {
-      name: 'Cyre Jorin To',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/cyrejorin/',
-      imageURL: cyre,
-    },
-    {
-      name: 'Dang Nguyen',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/dang-nguyen-nguyen-39b481223/',
-      imageURL: dang,
-    },
-    {
-      name: 'David Lu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/david-lin-yi-lu/',
-      imageURL: david,
-    },
-    {
-      name: 'Derek Lam',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/derekblam/',
-      imageURL: derek,
-    },
-    {
-      name: 'Dingyi (Jaedyn) Chang',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/dingyi-chang/',
-      imageURL: dingyi,
-    },
-    {
-      name: 'Donghyun Daniel Koo',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/koo-donghyun/',
-      imageURL: daniel,
-    },
-    {
-      name: 'Duong Vu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/duong-vu-641a861b2/',
-      imageURL: duongvu,
-    },
-    {
-      name: 'Edward Wu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/edward-william-wu/',
-      imageURL: edwardwu,
-    },
-    {
-      name: 'Ethan So',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/ethansso/',
-      imageURL: ethanso,
-    },
-    {
-      name: 'Evan Servito',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/evan-servito/',
-      imageURL: evan,
-    },
-    {
-      name: 'Gurneet Cheema',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/jeff-cheema/',
-      imageURL: gurneetcheema,
-    },
-    {
-      name: 'Hang Cao',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/hang-m-cao/',
-      imageURL: hang,
-    },
-    {
-      name: 'Hannah Chang',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/hannahchang10',
-      imageURL: hannahc,
-    },
-    {
-      name: 'Hassan Al-Najjar',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/hassan-al-najjar-438184215/',
-      imageURL: hassan,
-    },
-    {
-      name: 'Hayden Powers',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/hmpowers/',
-      imageURL: haydenp,
-    },
-    {
-      name: 'Ivy Nguyen',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/ivy-nguyen-19b87b220/',
-      imageURL: ivyn,
-    },
-    {
-      name: 'Jacqueline Thai',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/jthai-1201/',
-      imageURL: jackie,
-    },
-    {
-      name: 'Jared Lai',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/jared-l-02aa8b126',
-      imageURL: jared,
-    },
-    {
-      name: 'Jenny Lee',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/jenny-lee-6630201b2',
-      imageURL: jenny,
-    },
-    {
-      name: 'Johnny Phan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/jphan345/',
-      imageURL: johnny,
-    },
-    {
-      name: 'Joshua Chan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/joshualchan/',
-      imageURL: joshuachan,
-    },
-    {
-      name: 'Jules Labador',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/juleslabador',
-      imageURL: jules,
-    },
-    {
-      name: 'Justin Liao',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/justin-liao-1949b3217/',
-      imageURL: justin,
-    },
-    {
-      name: 'Joshua Montellano',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/joshua-montellano-b753391a3/',
-      imageURL: joshuamontellano,
-    },
-    {
-      name: 'Kanu Chandra',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kanu-chandra-30310916b/',
-      imageURL: kanu,
-    },
-    {
-      name: 'Kasey Chuang',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kaseyychuang',
-      imageURL: kasey,
-    },
-    {
-      name: 'Katie Chow',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/katie-chow-24a489166/',
-      imageURL: katie,
-    },
-    {
-      name: 'Kavan Dharia',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kavan-dharia-803861199/',
-      imageURL: kavan,
-    },
-    {
-      name: 'Keagan Nguyen',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/keagan-nguyen-a425a2176/',
-      imageURL: keagan,
-    },
-    {
-      name: 'Kei Asakawa',
-      position: 'Developer',
-      linkedinURL: 'www.linkedin.com/in/kei-asakawa',
-      imageURL: kei,
-    },
-    {
-      name: 'Kenny Tran',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kenny-tran-3020a715a/',
-      imageURL: kenny,
-    },
-    {
-      name: 'Koji Hirai',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/koji-hirai-5b508414a/',
-      imageURL: kojihirai,
-    },
-    {
-      name: 'Kory Zhang',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kory-zhang-00b86718b/',
-      imageURL: kory,
-    },
-    {
-      name: 'Kyle Doliente',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kyle-doliente-877133225/',
-      imageURL: kyledoliente,
-    },
-    {
-      name: 'Kyle Pan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/kylerpan/',
-      imageURL: kyle,
-    },
-    {
-      name: 'Lee Bahir',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/lee-bahir/',
-      imageURL: lee,
-    },
-    {
-      name: 'Madhusudhan Sharma',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/madhuxsharma/',
-      imageURL: madhusudhan,
-    },
-    {
-      name: 'Matthew Wu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/matthew-caleb-wu/',
-      imageURL: matthew,
-    },
-    {
-      name: 'Maxim Alayev',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/maximalayev/',
-      imageURL: maxim,
-    },
-    {
-      name: 'Mazen Ghosn',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/mazen-ghosn/',
-      imageURL: mazenghosn,
-    },
-    {
-      name: 'Megan Voong',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/meganvoong/',
-      imageURL: megan,
-    },
-    {
-      name: 'Melayna Navarro',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/melayna-navarro/',
-      imageURL: melayna,
-    },
-    {
-      name: 'Meryl Marasigan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/meryl-marasigan/',
-      imageURL: meryl,
-    },
-    {
-      name: 'Mia Schroeder',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/mia-schroeder',
-      imageURL: mia,
-    },
-    {
-      name: 'Michael Pien',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/michaelpien/',
-      imageURL: michaelpien,
-    },
-    {
-      name: 'Michelle Lin',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/michellelinCA',
-      imageURL: michellelin,
-    },
-    {
-      name: 'Michelle Lin',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/michellelin1',
-      imageURL: michelleblin,
-    },
-    {
-      name: 'Mignon April Lee',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/mignon-april-lee/',
-      imageURL: mignonaprillee,
-    },
-    {
-      name: 'Noah Arsala',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/noah-arsala-0324911b3/',
-      imageURL: noah,
-    },
-    {
-      name: 'Olivia Ih',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/olivia-ih/',
-      imageURL: olivia,
-    },
-    {
-      name: 'Orviel Alzate',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/orviel-marie-alzate-42898a224/',
-      imageURL: orvielalzate,
-    },
-    {
-      name: 'Parsa Karamat',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/parsa-karamat/',
-      imageURL: parsa,
-    },
-    {
-      name: 'Pham',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/nicoleminhpham/',
-      imageURL: pham,
-    },
-    {
-      name: 'Phil Truong',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/philip-truong/',
-      imageURL: phil,
-    },
-    {
-      name: 'Phillip Chan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/chnphllp/',
-      imageURL: phillip,
-    },
-    {
-      name: 'Preston Young',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/preston-c-young/',
-      imageURL: preston,
-    },
-    {
-      name: 'Qiran Wang',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/%E5%90%AF%E7%84%B6-%E6%B1%AA-7191b8225/',
-      imageURL: qiranwang,
-    },
-    {
-      name: 'Rachel Villamor',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/rachel-villamor/',
-      imageURL: rachel,
-    },
-    {
-      name: 'Rayan Tighiouart',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/rayan-tighiouart-b36989224/',
-      imageURL: rayan,
-    },
-    {
-      name: 'Riley Champion',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/rileychampion/',
-      imageURL: rileychampion,
-    },
-    {
-      name: 'Riley Wong',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/riley-wong/',
-      imageURL: rileywong,
-    },
-    {
-      name: 'Rithvik Tirumala',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/rithvik-tirumala/',
-      imageURL: rithvikT,
-    },
-    {
-      name: 'Rochelle Nixon',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/rochellenixon/',
-      imageURL: rochelle,
-    },
-    {
-      name: 'Rong Mu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/rong-lynn-mu-028992180/',
-      imageURL: rongmu,
-    },
-    {
-      name: 'Ruslan Manoharan',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/ruslan-manoharan-3b18801a4/',
-      imageURL: ruslan,
-    },
-    {
-      name: 'Ryan Cho',
-      position: 'Developer',
-      linkedinURL: 'www.linkedin.com/in/ryanjhcho',
-      imageURL: ryancho,
-    },
-    {
-      name: 'Sam Der',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/sam-der',
-      imageURL: sam,
-    },
-    {
-      name: 'Sravan Tallapaka',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/sravan-tallapaka-514288225/',
-      imageURL: sravantallapaka,
-    },
-    {
-      name: 'Ted Kim',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/ted-kim-471573203/',
-      imageURL: ted,
-    },
-    {
-      name: 'Thien Nguyen',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/thiennguyen2001/',
-      imageURL: thien,
-    },
-    {
-      name: 'Thomas Neill',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/thomas-neill-038b65224/',
-      imageURL: thomas,
-    },
-    {
-      name: 'Tony Liu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/tonyliu02/',
-      imageURL: tony,
-    },
-    {
-      name: 'Xuan Vu',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/xuanbvu/',
-      imageURL: xuan,
-    },
-    {
-      name: 'Zoya Hajee',
-      position: 'Developer',
-      linkedinURL: 'https://www.linkedin.com/in/zoyahajee/',
-      imageURL: zoya,
-    },
-    {
-      name: 'Alec Chen',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/alecc3/',
-      imageURL: alec,
-    },
-    {
-      name: 'Allison Liu',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/hiallisonliu/',
-      imageURL: allisonliu,
-    },
-    {
-      name: 'Alyssa Harvey',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/alyssajharvey/',
-      imageURL: alyssa,
-    },
-    {
-      name: 'Angeline Pho',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/angeline-pho-19743319a/',
-      imageURL: angeline,
-    },
-    {
-      name: 'Ashley Yeh',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/ashley-yeh/',
-      imageURL: ashleyyeh,
-    },
-    {
-      name: 'Becky Dinh',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/becky-dinh-27b85018b/',
-      imageURL: becky,
-    },
-    {
-      name: 'Carlos Lim',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/carlosalim7/',
-      imageURL: carlos,
-    },
-    {
-      name: 'Christina Vu',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/christina-vu-493560164/',
-      imageURL: christina,
-    },
-    {
-      name: 'Erica Yi',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/ericakyi/',
-      imageURL: erica,
-    },
-    {
-      name: 'Fiona Wu',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/fiona-wu-0/',
-      imageURL: fiona,
-    },
-    {
-      name: 'Hannah Limary',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/hannah-limary/',
-      imageURL: hannah,
-    },
-    {
-      name: 'Isabel Pham',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/isabe1pham/',
-      imageURL: isabel,
-    },
-    {
-      name: 'Jacqueline Le',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/jacquelinexle/',
-      imageURL: jacqueline,
-    },
-    {
-      name: 'Jade Luu',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/jade-luu-883209203/',
-      imageURL: jade,
-    },
-    {
-      name: 'James Molamphy',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/james-molamphy/',
-      imageURL: james,
-    },
-    {
-      name: 'Juhi Patel',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/juhipatel21/',
-      imageURL: juhi,
-    },
-    {
-      name: 'Kaila Long',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/kaila-long-3952b7210/',
-      imageURL: kaila,
-    },
-    {
-      name: 'Kaitlyn Suzuki',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/kaitlyn-suzuki/',
-      imageURL: kaitlyn,
-    },
-    {
-      name: 'Lancy Tan',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/lancy-tan-b79885142/',
-      imageURL: lancy,
-    },
-    {
-      name: 'Maireen Espiritu',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/maireen-espiritu/',
-      imageURL: maireen,
-    },
-    {
-      name: 'Michelle Huynh',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/michelleyhuynh/',
-      imageURL: michelle,
-    },
-    {
-      name: 'Minh Nguyen',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/minh-nguyen16/',
-      imageURL: minh,
-    },
-    {
-      name: 'Serena Trang',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/serenatrang/',
-      imageURL: serena,
-    },
-    {
-      name: 'Vani Anilkumar',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/vani-anilkumar-670819137/',
-      imageURL: vani,
-    },
-    {
-      name: 'Zoya Soy',
-      position: 'Designer',
-      linkedinURL: 'https://www.linkedin.com/in/zoyasoy/',
-      imageURL: zoyasoy,
-    },
-  ],
-};
-
-export default members;
