+++ website/src/app/views/our-team/members.js
@@ -1,0 +1,91 @@
+import jane from '../../../images/about/memberImages/jane.jpg';
+import maxim from '../../../images/about/memberImages/maxim.jpeg';
+import meghna from '../../../images/about/memberImages/meghna.jpg';
+import alan from '../../../images/about/memberImages/alan.jpg';
+import claude from '../../../images/about/memberImages/claude.jpg';
+import alexandria from '../../../images/about/memberImages/alexandria.jpg';
+import brie from '../../../images/about/memberImages/brie.jpg';
+import hang from '../../../images/about/memberImages/hang.jpg';
+import justin from '../../../images/about/memberImages/justin.jpg';
+import zoya from '../../../images/about/memberImages/zoyasoy.png';
+
+const members = {
+    board: {
+        name: "Board",
+        members: [
+            {
+                name: 'Jane Vo',
+                position: 'President',
+                linkedinURL: 'https://www.linkedin.com/in/janehv/',
+                imageURL: jane,
+            },
+            {
+                name: 'Maxim Alayev',
+                position: 'Internal Vice President',
+                linkedinURL: 'https://www.linkedin.com/in/maximalayev/',
+                imageURL: maxim,
+            },
+            {
+                name: 'Meghna Kaligotla',
+                position: 'External Vice President',
+                linkedinURL: 'https://www.linkedin.com/in/meghna-kaligotla-605/',
+                imageURL: meghna,
+            },
+            {
+                name: 'Alan Chang',
+                position: 'Co-Director of Technical',
+                linkedinURL: 'https://www.linkedin.com/in/acxyz/',
+                imageURL: alan,
+            },
+            {
+                name: 'Claude Yan',
+                position: 'Co-Director of Technical',
+                linkedinURL: 'https://www.linkedin.com/in/claude-yan-b41a3216a/',
+                imageURL: claude,
+            },
+            {
+                name: 'Sabrina Nguyen-Hoang',
+                position: 'Co-Director of Design',
+                linkedinURL: 'https://www.linkedin.com/in/sabrina-nguyen-hoang/',
+                imageURL: brie,
+            },
+            {
+                name: 'Alexandria Wang',
+                position: 'Co-Director of Design',
+                linkedinURL: 'https://www.linkedin.com/in/alexandria-wang-44066a187/',
+                imageURL: alexandria,
+            },
+        ],
+    },
+    npo1: {
+        name: "NPO 1",
+        members: [
+            {
+                name: 'Hang Cao',
+                position: 'Tech Lead',
+                linkedinURL: 'https://www.linkedin.com/in/hang-m-cao/',
+                imageURL: hang,
+            },
+            {
+                name: 'Justin Liao',
+                position: 'Tech Lead',
+                linkedinURL: 'https://www.linkedin.com/in/justin-liao-1949b3217/',
+                imageURL: justin,
+            },
+            {
+                name: 'Zoya Soy',
+                position: 'Designer',
+                linkedinURL: 'https://www.linkedin.com/in/zoyasoy/',
+                imageURL: zoya,
+            },
+        ],
+    },
+    npo2: {
+
+    },
+    npo3: {
+        
+    }
+};
+
+export default members;
